+++ src/controllers/CkeditorController.php
@@ -1,0 +1,46 @@
+<?php
+
+namespace craft\ckeditor\controllers;
+
+use craft\elements\Asset;
+use craft\web\Controller;
+use yii\web\NotFoundHttpException;
+use yii\web\Response;
+
+/**
+ * CKEditor controller
+ *
+ * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
+ * @since 3.4.0
+ */
+class CkeditorController extends Controller
+{
+    public function beforeAction($action): bool
+    {
+        if (!parent::beforeAction($action)) {
+            return false;
+        }
+
+        $this->requireCpRequest();
+        return true;
+    }
+
+    public function actionImageUrl(): Response
+    {
+        $assetId = $this->request->getRequiredBodyParam('assetId');
+        $transform = $this->request->getBodyParam('transform');
+
+        $asset = Asset::find()
+            ->id($assetId)
+            ->kind('image')
+            ->one();
+
+        if (!$asset) {
+            throw new NotFoundHttpException('Image not found');
+        }
+
+        return $this->asJson([
+            'url' => $asset->getUrl($transform, false),
+        ]);
+    }
+}